password reset config reorganized
