--- backend/src/main/scala/com/softwaremill/bootzooka/email/EmailContentWithSubject.scala
@@ -1,3 +1,0 @@
-package com.softwaremill.bootzooka.email
-
-case class EmailContentWithSubject(content: String, subject: String)
