--- backend/src/main/scala/com/softwaremill/bootzooka/email/EmailService.scala
@@ -1,9 +1,0 @@
-package com.softwaremill.bootzooka.email
-
-import scala.concurrent.Future
-
-trait EmailService {
-
-  def scheduleEmail(address: String, emailData: EmailContentWithSubject): Future[Unit]
-
-}
