+++ backend/src/main/scala/com/softwaremill/bootzooka/email/application/DummyEmailService.scala
@@ -1,6 +1,15 @@
-package com.softwaremill.bootzooka.email
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
-import SmtpEmailSender.EmailDescription
+package com.softwaremill.bootzooka.email.application
+
+import com.softwaremill.bootzooka.email.application.SmtpEmailSender.EmailDescription
+import com.softwaremill.bootzooka.email.domain.EmailContentWithSubject
 import com.typesafe.scalalogging.StrictLogging
 
 import scala.collection.mutable.ListBuffer
