+++ backend/src/main/scala/com/softwaremill/bootzooka/email/application/SmtpEmailService.scala
@@ -1,9 +1,18 @@
-package com.softwaremill.bootzooka.email
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
+package com.softwaremill.bootzooka.email.application
+
+import com.softwaremill.bootzooka.email.domain.EmailContentWithSubject
 import com.typesafe.scalalogging.StrictLogging
 
 import scala.concurrent._
-import scala.util.{Failure, Success, Try}
+import scala.util.{Failure, Success}
 
 class SmtpEmailService(emailConfig: EmailConfig)(implicit ec: ExecutionContext) extends EmailService with StrictLogging {
   def scheduleEmail(address: String, email: EmailContentWithSubject) = {
