+++ backend/src/test/scala/com/softwaremill/bootzooka/test/TestHelpersWithDb.scala
@@ -1,11 +1,9 @@
 package com.softwaremill.bootzooka.test
 
-import com.softwaremill.bootzooka.CoreConfig
 import com.softwaremill.bootzooka.common.sql.SqlDatabase
-import com.softwaremill.bootzooka.email.{DummyEmailService, EmailTemplatingEngine}
-import com.softwaremill.bootzooka.user.domain.User
+import com.softwaremill.bootzooka.email.application.{DummyEmailService, EmailTemplatingEngine}
 import com.softwaremill.bootzooka.user.application.{UserDao, UserService}
-import com.typesafe.config.ConfigFactory
+import com.softwaremill.bootzooka.user.domain.User
 import org.scalatest.concurrent.ScalaFutures
 
 import scala.concurrent.ExecutionContext
@@ -16,11 +14,9 @@ trait TestHelpersWithDb extends TestHelpers with ScalaFutures {
   lazy val emailTemplatingEngine = new EmailTemplatingEngine
   lazy val userDao = new UserDao(sqlDatabase)
   lazy val userService = new UserService(userDao, emailService, emailTemplatingEngine)
-  lazy val config = new CoreConfig {
-    override def rootConfig = ConfigFactory.load()
-  }
 
   def sqlDatabase: SqlDatabase
+
   implicit lazy val ec: ExecutionContext = scala.concurrent.ExecutionContext.Implicits.global
 
   def newRandomStoredUser(password: Option[String] = None): User = {
