+++ ui-tests/src/test/scala/uitest/BaseUiSpec.scala
@@ -3,7 +3,7 @@ package uitest
 import java.util.concurrent.TimeUnit
 
 import akka.http.scaladsl.Http.ServerBinding
-import com.softwaremill.bootzooka.email.DummyEmailService
+import com.softwaremill.bootzooka.email.application.DummyEmailService
 import com.softwaremill.bootzooka.passwordreset.application.SqlPasswordResetCodeSchema
 import com.softwaremill.bootzooka.user.application.SqlUserSchema
 import com.softwaremill.bootzooka.{BusinessLogic, Main}
